requirements.txt
Pillow==9.2.0
PyInstaller==5.4
tkinterdnd2==0.3.0
